Disallow IPv4 addresses with leading 0s in the quads
This is a security issue for other modules such as Net::CIDR::Lite,
where the leading zero causes the address to be treated as octal.

Although the issue does not seem to affect this module, the code comes
from Net::CIDR::Lite so the corresponding fix has been incorporated as a
precaution.